fixed ai

+++ game/server/ai_behavior_follow.cpp
@@ -53,7 +53,7 @@ struct AI_Follower_t
 	}
 
 	AIHANDLE 			hFollower;
-	int					slot;
+	intp				slot;
 	AI_FollowNavInfo_t	navInfo;
 	AI_FollowGroup_t *	pGroup;	// backpointer for efficiency
 };
@@ -2641,8 +2641,8 @@ bool CAI_FollowManager::RedistributeSlots( AI_FollowGroup_t *pGroup )
 	{
 		AI_FollowSlot_t *  pSlot 	  = &pGroup->pFormation->pSlots[bestSlot];
 		Vector			   slotPos	  = originFollowed + pSlot->position;
-		int  h			= pGroup->followers.Head();
-		int  hBest 		= pGroup->followers.InvalidIndex();
+		intp  h			   = pGroup->followers.Head();
+		intp  hBest 	   = pGroup->followers.InvalidIndex();
 		float 			   distSqBest = FLT_MAX;
 		
 		while ( h != pGroup->followers.InvalidIndex() )
@@ -2691,7 +2691,7 @@ void CAI_FollowManager::ChangeFormation( AI_FollowManagerInfoHandle_t& hInfo, AI
 	if ( pNewFormation == pGroup->pFormation )
 		return;
 
-	int h = pGroup->followers.Head();
+	intp h = pGroup->followers.Head();
 		
 	while ( h != pGroup->followers.InvalidIndex() )
 	{
@@ -2738,7 +2738,7 @@ void CAI_FollowManager::RemoveFollower( AI_FollowManagerInfoHandle_t& hInfo )
 		AI_FollowGroup_t *pGroup = hInfo.m_pGroup;
 		AI_Follower_t* iterNode = &pGroup->followers[hInfo.m_hFollower];
 
-		int slot = iterNode->slot;
+		intp slot = iterNode->slot;
 		pGroup->slotUsage.Clear( slot );
 		pGroup->followers.Remove( hInfo.m_hFollower );
 		if ( pGroup->followers.Count() == 0 )
